Resolve some conversion warnings

+++ lib/pathplan/cvt.c
@@ -118,7 +118,7 @@ int Pobspath(vconfig_t * config, Ppoint_t p0, int poly0, Ppoint_t p1,
 	     int poly1, Ppolyline_t * output_route)
 {
     int i, j, *dad;
-    int opn;
+    size_t opn;
     Ppoint_t *ops;
     COORD *ptvis0, *ptvis1;
